// cosmowcl/src/app/shared/models/api/cosmoauthorization/certificatoFirma.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * impostazioni-firma
 * API di configurazione impostazioni firma
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TipoOTP } from './tipoOTP.d';
import { ProfiloFEQ } from './profiloFEQ.d';
import { SceltaMarcaTemporale } from './sceltaMarcaTemporale.d';
import { EnteCertificatore } from './enteCertificatore.d';
import { TipoCredenzialiFirma } from './tipoCredenzialiFirma.d';


/**
 * 
 */
export interface CertificatoFirma { 
    id?: number;
    descrizione: string;
    username: string;
    pin: string;
    dataScadenza?: string;
    enteCertificatore: EnteCertificatore;
    profiloFEQ: ProfiloFEQ;
    sceltaMarcaTemporale: SceltaMarcaTemporale;
    tipoCredenzialiFirma: TipoCredenzialiFirma;
    tipoOTP: TipoOTP;
    password?: string;
    ultimoUtilizzato?: boolean;
}